1) Implement the getDoubleValue and validateDouble in the BaseActionForm 2) Refactor code to use getDoubleValue instead of Double.parseDouble or Double.valueOf 3) Fix code to use Money instead of double for calculation 4) Implement getDoubleValue for money format and interest format. 5) Configure money format and interest format in the application configuration file and the AccoutingRule will get the configured values. 6) Change the ranges of loans to integers instead of double 7) Move error messages to properties file from LoanPrdActionForm 8) Validate users' inputs for double in LoanPrdActionForm (the checking for valid numbers entered still done in javascript)
git-svn-id: https://mifos.svn.sourceforge.net/svnroot/mifos/trunk@12242 a8845c50-7012-0410-95d3-8e1449b9b1e4

+++ mifos/test/org/mifos/application/accounts/loan/business/TestLoanBO.java
@@ -5102,7 +5102,7 @@ public class TestLoanBO extends MifosTestCase {
 		TestObjectFactory.flushandCloseSession();
 		accountBO = TestObjectFactory.getObject(AccountBO.class, accountBO
 				.getAccountId());
-		assertEquals((float) 1.0, ((LoanBO) accountBO).getPaymentsInArrears());
+		assertEquals((double) 1.0, ((LoanBO) accountBO).getPaymentsInArrears());
 
 	}
